requirements.txt
altgraph==0.17.4
certifi==2024.7.4
charset-normalizer==3.3.2
colorthief==0.2.1
contourpy==1.2.1
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.53.1
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.1.post1
numpy==2.0.1
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
plyer==2.1.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.2
PyQt6==6.7.1
PyQt6-Fluent-Widgets==1.6.0
PyQt6-Frameless-Window==0.3.9
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
pywin32==306
pywin32-ctypes==0.2.2
requests==2.32.3
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
urllib3==2.2.2
